requirements.txt
networkx==3.2.1
matplotlib==3.8.2
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.1
osmium==3.7.0
folium==0.15.1 